requirements.txt
requests
pyobjc
PyInquirer